third_party/WebKit/LayoutTests/bluetooth/idl-BluetoothUuids.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_throws(null, function() { new BluetoothUuids(); },
                'the constructor should not be callable with "new"');
  assert_throws(null, function() { BluetoothUuids(); },
                'the constructor should not be callable');
}, 'BluetoothUuids IDL test');
</script>
